Float64 time/fraction precision: follow-path times, uniform-dt gate, draw_ranges/clip_times/point_times channels (#100)
* Float64 time vector for follow-path tracks + cumulative uniform-dt gate

Two length-scaling time-precision fixes for long-toolpath playback:

- set_follow_path now ships times as (K,) float64 followed by (K,6)
  float32 [px,py,pz,ax,ay,az] (was float32 rows of 7). Float32's ulp is
  ~4 ms at t=40,000 s and ~15.6 ms at 160,000 s, which wobbled the tool
  velocity per segment and collapsed ms-spaced keys to dt=0
  (hold-then-jump stutter) late in hours-long timelines.
- The uniform-dt fast-path gate now bounds CUMULATIVE deviation from
  t0 + i*dt0 instead of the per-interval delta: the fast path derives
  the frame index from dt0 alone, so per-interval tolerance let error
  accumulate linearly (tens of frames of drift on a 40k-frame cumsum
  timeline), desyncing keyframed objects from follow-path tracks.


* Float64 animation channels: draw_ranges, clip_times, point_times

Second float32 vector on long toolpaths: draw_ranges values are
fractions of TOTAL arc length, so one float32 ulp near 1.0 (~6e-8) is
~0.5 mm of frontier position on a multi-km spine — keyframe knots
wobble against the follow-path tool. clip_times/point_times are
absolute seconds with the same magnitude problem (~16 ms ulp at
t=160,000 s; ms-spaced knots collapse to equal values).

- add_channel accepts dtype="float64"; the three convenience setters
  pack it by default. transforms/camera channels stay float32.
- Packer sorts channels by element size descending so the float64
  block leads the payload (Float64Array views need 8-byte alignment);
  the viewer copies misaligned hand-packed handleMessage payloads
  instead of throwing.


* Remove the issue #97 EMA pace smoothing; raw wall-clock advance + stall cap

The EMA papered over playhead jitter whose actual source was float32
time/fraction quantization, now fixed at the source (this PR). It also
added its own artefact: the playhead lags/leads wall time while the
average catches up. Playback now advances by the raw per-rAF wall-clock
delta — real time is the pacing reference — with a single-delta cap
(PLAYBACK_MAX_FRAME_DELTA = 0.25 s) so a backgrounded tab or GC stall
can't teleport the playhead by minutes x speed.


---------

+++ tests/test_animation.py
@@ -113,7 +113,7 @@ def test_set_draw_range_data():
 
     assert len(animation._channels) == 1
     assert animation._channels[0].name == "draw_ranges"
-    assert animation._channels[0].dtype == "float32"
+    assert animation._channels[0].dtype == "float64"
     assert animation._channels[0].stride == 1
 
 
@@ -128,7 +128,7 @@ def test_set_clip_time_data():
 
     assert len(animation._channels) == 1
     assert animation._channels[0].name == "clip_times"
-    assert animation._channels[0].dtype == "float32"
+    assert animation._channels[0].dtype == "float64"
     assert animation._channels[0].stride == 1
 
 
@@ -143,7 +143,7 @@ def test_set_point_time_data():
 
     assert len(animation._channels) == 1
     assert animation._channels[0].name == "point_times"
-    assert animation._channels[0].dtype == "float32"
+    assert animation._channels[0].dtype == "float64"
     assert animation._channels[0].stride == 1
     assert animation._channels[0].interpolation == "linear"
